+++ packages/auth/__tests__/hosted-ui.test.ts
@@ -300,11 +300,9 @@ describe('Hosted UI tests', () => {
 
 		expect.assertions(2);
 
-		try {
-			await auth.signOut({ global: true });
-		} catch (err) {
-			expect(err).toEqual('Signout timeout fail');
-		}
+		await expect(auth.signOut({ global: true })).rejects.toThrowError(
+			'Signout timeout fail'
+		);
 
 		expect(spyGlobalSignOut).toBeCalled();
 	});
@@ -450,11 +448,9 @@ describe('Hosted UI tests', () => {
 
 		expect.assertions(1);
 
-		try {
-			await auth.signOut({ global: false });
-		} catch (err) {
-			expect(err).toEqual('Signout timeout fail');
-		}
+		await expect(auth.signOut({ global: false })).rejects.toThrowError(
+			'Signout timeout fail'
+		);
 	});
 
 	test('globalSignOut hosted ui, url opener', done => {
@@ -519,17 +515,15 @@ describe('Hosted UI tests', () => {
 			};
 		});
 
-		const urlOpener = jest.fn(
-			(url: string): Promise<any> => {
-				return new Promise(() => {
-					expect(url).toEqual(
-						'https://xxxxxxxxxxxx-xxxxxx-xxx.auth.us-west-2.amazoncognito.com/logout?client_id=awsUserPoolsWebClientId&logout_uri=http%3A%2F%2Flocalhost%3A4200%2F'
-					);
+		const urlOpener = jest.fn((url: string): Promise<any> => {
+			return new Promise(() => {
+				expect(url).toEqual(
+					'https://xxxxxxxxxxxx-xxxxxx-xxx.auth.us-west-2.amazoncognito.com/logout?client_id=awsUserPoolsWebClientId&logout_uri=http%3A%2F%2Flocalhost%3A4200%2F'
+				);
 
-					done();
-				});
-			}
-		);
+				done();
+			});
+		});
 		const options = {
 			...authOptionsWithOAuth,
 		};