+++ packages/datastore/__tests__/mutation.test.ts
@@ -1,4 +1,7 @@
-import { MutationProcessor } from '../src/sync/processors/mutation';
+import {
+	MutationProcessor,
+	safeJitteredBackoff,
+} from '../src/sync/processors/mutation';
 import {
 	Model as ModelType,
 	PostCustomPK as PostCustomPKType,
@@ -20,6 +23,53 @@ let Model: PersistentModelConstructor<ModelType>;
 let PostCustomPK: PersistentModelConstructor<PostCustomPKType>;
 let PostCustomPKSort: PersistentModelConstructor<PostCustomPKSortType>;
 
+describe('Jittered retry', () => {
+	it('should progress exponentially until some limit', () => {
+		const COUNT = 13;
+
+		const backoffs = [...Array(COUNT)].map((v, i) =>
+			safeJitteredBackoff(i)
+		) as (number | boolean)[];
+
+		const isExpectedValue = (value, attempt) => {
+			const lowerLimit = 2 ** attempt * 100;
+			const upperLimit = lowerLimit + 100;
+
+			if (lowerLimit < 2 ** 12 * 100) {
+				console.log(
+					`attempt ${attempt} (${value}) should be between ${lowerLimit} and ${upperLimit} inclusively.`
+				);
+				return value >= lowerLimit && value <= upperLimit;
+			} else {
+				console.log(`attempt ${attempt} (${value}) should be false.`);
+				return value === false;
+			}
+		};
+
+		backoffs.forEach((value, attempt) => {
+			expect(isExpectedValue(value, attempt)).toBe(true);
+		});
+
+		// we should be testing up to the edge. at least one backoff at the
+		// end of the list must be false. (past the limit)
+		expect(backoffs.pop()).toBe(false);
+	});
+
+	it('should retry forever on network errors', () => {
+		const MAX_DELAY = 5 * 60 * 1000;
+		const COUNT = 1000;
+
+		const backoffs = [...Array(COUNT)].map((v, i) =>
+			safeJitteredBackoff(i, [], new Error('Network Error'))
+		) as (number | boolean)[];
+
+		backoffs.forEach(v => {
+			expect(v).toBeTruthy();
+			expect(v).toBeLessThanOrEqual(MAX_DELAY);
+		});
+	});
+});
+
 describe('MutationProcessor', () => {
 	let mutationProcessor: MutationProcessor;
 
@@ -34,11 +84,11 @@ describe('MutationProcessor', () => {
 
 			await mutationProcessor.resume();
 
-			expect(mockJitteredExponentialRetry.mock.results).toHaveLength(1);
+			expect(mockRetry.mock.results).toHaveLength(1);
 
-			await expect(
-				mockJitteredExponentialRetry.mock.results[0].value
-			).rejects.toEqual(new Error('Network Error'));
+			await expect(mockRetry.mock.results[0].value).rejects.toEqual(
+				new Error('Network Error')
+			);
 
 			expect(mutationProcessorSpy).toHaveBeenCalled();
 
@@ -137,19 +187,17 @@ jest.mock('@aws-amplify/api', () => {
 	};
 });
 
-// mocking jitteredExponentialRetry to prevent it from retrying
+// mocking jitteredBackoff to prevent it from retrying
 // endlessly in the mutation processor and so that we can expect the thrown result in our test
 // should throw a Network Error
-let mockJitteredExponentialRetry;
+let mockRetry;
 jest.mock('@aws-amplify/core', () => {
-	mockJitteredExponentialRetry = jest
-		.fn()
-		.mockImplementation(async (fn, args) => {
-			await fn(...args);
-		});
+	mockRetry = jest.fn().mockImplementation(async (fn, args) => {
+		await fn(...args);
+	});
 	return {
 		...jest.requireActual('@aws-amplify/core'),
-		jitteredExponentialRetry: mockJitteredExponentialRetry,
+		retry: mockRetry,
 	};
 });
 
@@ -220,6 +268,7 @@ async function instantiateMutationProcessor() {
 			aws_appsync_authenticationType: 'API_KEY',
 			aws_appsync_apiKey: 'da2-xxxxxxxxxxxxxxxxxxxxxx',
 		},
+		() => null,
 		() => null
 	);