+++ packages/datastore/src/sync/index.ts
@@ -146,13 +146,15 @@ export class SyncEngine {
 			this.schema,
 			this.syncPredicates,
 			this.amplifyConfig,
-			this.authModeStrategy
+			this.authModeStrategy,
+			errorHandler
 		);
 		this.subscriptionsProcessor = new SubscriptionProcessor(
 			this.schema,
 			this.syncPredicates,
 			this.amplifyConfig,
-			this.authModeStrategy
+			this.authModeStrategy,
+			errorHandler
 		);
 		this.mutationsProcessor = new MutationProcessor(
 			this.schema,
@@ -163,8 +165,8 @@ export class SyncEngine {
 			MutationEvent,
 			this.amplifyConfig,
 			this.authModeStrategy,
-			conflictHandler,
-			errorHandler
+			errorHandler,
+			conflictHandler
 		);
 		this.datastoreConnectivity = new DataStoreConnectivity();
 	}