+++ packages/datastore/src/sync/merger.ts
@@ -7,6 +7,8 @@ import {
 } from '../types';
 import { MutationEventOutbox } from './outbox';
 import { getIdentifierValue } from './utils';
+
+// https://github.com/aws-amplify/amplify-js/blob/datastore-docs/packages/datastore/docs/sync-engine.md#merger
 class ModelMerger {
 	constructor(
 		private readonly outbox: MutationEventOutbox,