+++ packages/geo/__tests__/Geo.test.ts
@@ -13,9 +13,8 @@
 import { Credentials } from '@aws-amplify/core';
 import {
 	LocationClient,
-	SearchPlaceIndexForTextCommand,
-	SearchPlaceIndexForSuggestionsCommand,
 	SearchPlaceIndexForPositionCommand,
+	SearchPlaceIndexForTextCommand,
 } from '@aws-sdk/client-location';
 
 import { GeoClass } from '../src/Geo';
@@ -31,7 +30,18 @@ import {
 	awsConfig,
 	TestPlacePascalCase,
 	testPlaceCamelCase,
-} from './data';
+	validGeometry,
+	validGeofences,
+	validGeofence1,
+	singleGeofenceCamelcaseResults,
+	batchGeofencesCamelcaseResults,
+} from './testData';
+
+import {
+	mockBatchPutGeofenceCommand,
+	mockGetGeofenceCommand,
+	mockListGeofencesCommand,
+} from './testUtils';
 
 LocationClient.prototype.send = jest.fn(async command => {
 	if (
@@ -42,18 +52,6 @@ LocationClient.prototype.send = jest.fn(async command => {
 			Results: [
 				{
 					Place: TestPlacePascalCase,
-				},
-			],
-		};
-	}
-	if (command instanceof SearchPlaceIndexForSuggestionsCommand) {
-		return {
-			Results: [
-				{
-					Text: 'star',
-				},
-				{
-					Text: 'not star',
 				},
 			],
 		};
@@ -147,7 +145,7 @@ describe('Geo', () => {
 			geo.configure({});
 
 			expect(() => geo.getAvailableMaps()).toThrow(
-				"No map resources found in amplify config, run 'amplify add geo' to create them and run `amplify push` after"
+				"No map resources found in amplify config, run 'amplify add geo' to create one and run `amplify push` after"
 			);
 		});
 
@@ -172,7 +170,7 @@ describe('Geo', () => {
 			geo.configure({});
 
 			expect(() => geo.getDefaultMap()).toThrow(
-				"No map resources found in amplify config, run 'amplify add geo' to create them and run `amplify push` after"
+				"No map resources found in amplify config, run 'amplify add geo' to create one and run `amplify push` after"
 			);
 		});
 
@@ -207,7 +205,7 @@ describe('Geo', () => {
 	});
 
 	describe('searchByText', () => {
-		const testString = 'star';
+		const testString = 'starbucks';
 
 		test('should search with just text input', async () => {
 			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
@@ -314,131 +312,15 @@ describe('Geo', () => {
 			geo.configure(awsConfig);
 			geo.removePluggable('AmazonLocationService');
 
+			const testString = 'starbucks';
 			await expect(geo.searchByText(testString)).rejects.toThrow(
 				'No plugin found in Geo for the provider'
 			);
 		});
 	});
 
-	describe('searchForSuggestions', () => {
-		const testString = 'star';
-		const testResults = ['star', 'not star'];
-
-		test('should search with just text input', async () => {
-			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
-				return Promise.resolve(credentials);
-			});
-
-			const geo = new GeoClass();
-			geo.configure(awsConfig);
-
-			const results = await geo.searchForSuggestions(testString);
-			expect(results).toEqual(testResults);
-
-			const spyon = jest.spyOn(LocationClient.prototype, 'send');
-			const input = spyon.mock.calls[0][0].input;
-			expect(input).toEqual({
-				Text: testString,
-				IndexName: awsConfig.geo.amazon_location_service.search_indices.default,
-			});
-		});
-
-		test('should search using given options with biasPosition', async () => {
-			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
-				return Promise.resolve(credentials);
-			});
-
-			const geo = new GeoClass();
-			geo.configure(awsConfig);
-
-			const searchOptions: SearchByTextOptions = {
-				biasPosition: [12345, 67890],
-				countries: ['USA'],
-				maxResults: 40,
-				searchIndexName: 'geoJSSearchCustomExample',
-			};
-			const results = await geo.searchForSuggestions(testString, searchOptions);
-			expect(results).toEqual(testResults);
-
-			const spyon = jest.spyOn(LocationClient.prototype, 'send');
-			const input = spyon.mock.calls[0][0].input;
-			expect(input).toEqual({
-				Text: testString,
-				IndexName: searchOptions.searchIndexName,
-				BiasPosition: searchOptions.biasPosition,
-				FilterCountries: searchOptions.countries,
-				MaxResults: searchOptions.maxResults,
-			});
-		});
-
-		test('should search using given options with searchAreaConstraints', async () => {
-			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
-				return Promise.resolve(credentials);
-			});
-
-			const geo = new GeoClass();
-			geo.configure(awsConfig);
-
-			const searchOptions: SearchByTextOptions = {
-				searchAreaConstraints: [123, 456, 789, 321],
-				countries: ['USA'],
-				maxResults: 40,
-				searchIndexName: 'geoJSSearchCustomExample',
-			};
-			const results = await geo.searchForSuggestions(testString, searchOptions);
-			expect(results).toEqual(testResults);
-
-			const spyon = jest.spyOn(LocationClient.prototype, 'send');
-			const input = spyon.mock.calls[0][0].input;
-			expect(input).toEqual({
-				Text: testString,
-				IndexName: searchOptions.searchIndexName,
-				FilterBBox: searchOptions.searchAreaConstraints,
-				FilterCountries: searchOptions.countries,
-				MaxResults: searchOptions.maxResults,
-			});
-		});
-
-		test('should throw an error if both BiasPosition and SearchAreaConstraints are given in the options', async () => {
-			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
-				return Promise.resolve(credentials);
-			});
-
-			const geo = new GeoClass();
-			geo.configure(awsConfig);
-
-			const searchOptions: SearchByTextOptions = {
-				countries: ['USA'],
-				maxResults: 40,
-				searchIndexName: 'geoJSSearchCustomExample',
-				biasPosition: [12345, 67890],
-				searchAreaConstraints: [123, 456, 789, 321],
-			};
-
-			await expect(
-				geo.searchForSuggestions(testString, searchOptions)
-			).rejects.toThrow(
-				'BiasPosition and SearchAreaConstraints are mutually exclusive, please remove one or the other from the options object'
-			);
-		});
-
-		test('should fail if there is no provider', async () => {
-			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
-				return Promise.resolve(credentials);
-			});
-
-			const geo = new GeoClass();
-			geo.configure(awsConfig);
-			geo.removePluggable('AmazonLocationService');
-
-			await expect(geo.searchForSuggestions(testString)).rejects.toThrow(
-				'No plugin found in Geo for the provider'
-			);
-		});
-	});
-
 	describe('searchByCoordinates', () => {
-		const testCoordinates: Coordinates = [12345, 67890];
+		const testCoordinates: Coordinates = [45, 90];
 
 		test('should search with just coordinate input', async () => {
 			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
@@ -497,6 +379,161 @@ describe('Geo', () => {
 
 			await expect(geo.searchByCoordinates(testCoordinates)).rejects.toThrow(
 				'No plugin found in Geo for the provider'
+			);
+		});
+	});
+
+	describe('saveGeofences', () => {
+		test('saveGeofences with a single geofence', async () => {
+			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
+				return Promise.resolve(credentials);
+			});
+
+			LocationClient.prototype.send = jest
+				.fn()
+				.mockImplementationOnce(mockBatchPutGeofenceCommand);
+
+			const geo = new GeoClass();
+			geo.configure(awsConfig);
+
+			// Check that results are what's expected
+			const results = await geo.saveGeofences(validGeofence1);
+			expect(results).toEqual(singleGeofenceCamelcaseResults);
+
+			// Expect that the API was called with the proper input
+			const spyon = jest.spyOn(LocationClient.prototype, 'send');
+			const input = spyon.mock.calls[0][0].input;
+			const output = {
+				Entries: [
+					{
+						GeofenceId: validGeofence1.geofenceId,
+						Geometry: {
+							Polygon: validGeofence1.geometry.polygon,
+						},
+					},
+				],
+				CollectionName: 'geofenceCollectionExample',
+			};
+			expect(input).toEqual(output);
+		});
+
+		test('saveGeofences with multiple geofences', async () => {
+			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
+				return Promise.resolve(credentials);
+			});
+
+			LocationClient.prototype.send = jest
+				.fn()
+				.mockImplementation(mockBatchPutGeofenceCommand);
+
+			const geo = new GeoClass();
+			geo.configure(awsConfig);
+
+			// Check that results are what's expected
+			const results = await geo.saveGeofences(validGeofences);
+			expect(results).toEqual(batchGeofencesCamelcaseResults);
+
+			// Expect that the API was called the right amount of times
+			const expectedNumberOfCalls = Math.floor(validGeofences.length / 10) + 1;
+			expect(LocationClient.prototype.send).toHaveBeenCalledTimes(
+				expectedNumberOfCalls
+			);
+		});
+
+		test('should fail if there is no provider', async () => {
+			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
+				return Promise.resolve(credentials);
+			});
+
+			const geo = new GeoClass();
+			geo.configure(awsConfig);
+			geo.removePluggable('AmazonLocationService');
+
+			await expect(geo.saveGeofences(validGeofence1)).rejects.toThrow(
+				'No plugin found in Geo for the provider'
+			);
+		});
+	});
+
+	describe('getGeofence', () => {
+		test('getGeofence returns the right geofence', async () => {
+			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
+				return Promise.resolve(credentials);
+			});
+
+			LocationClient.prototype.send = jest
+				.fn()
+				.mockImplementationOnce(mockGetGeofenceCommand);
+
+			const geo = new GeoClass();
+			geo.configure(awsConfig);
+
+			// Check that results are what's expected
+			const results = await geo.getGeofence('testGeofenceId');
+			const expected = {
+				geofenceId: 'testGeofenceId',
+				geometry: validGeometry,
+				createTime: '2020-04-01T21:00:00.000Z',
+				updateTime: '2020-04-01T21:00:00.000Z',
+				status: 'ACTIVE',
+			};
+			expect(results).toEqual(expected);
+
+			// Expect that the API was called with the proper input
+			const spyon = jest.spyOn(LocationClient.prototype, 'send');
+			const input = spyon.mock.calls[0][0].input;
+			const output = {
+				GeofenceId: 'testGeofenceId',
+				CollectionName: 'geofenceCollectionExample',
+			};
+			expect(input).toEqual(output);
+		});
+	});
+
+	describe('listGeofences', () => {
+		test('listGeofences gets the first 100 geofences when no arguments are given', async () => {
+			jest.spyOn(Credentials, 'get').mockImplementationOnce(() => {
+				return Promise.resolve(credentials);
+			});
+
+			LocationClient.prototype.send = jest
+				.fn()
+				.mockImplementationOnce(mockListGeofencesCommand);
+
+			const geo = new GeoClass();
+			geo.configure(awsConfig);
+
+			// Check that results are what's expected
+			const results = await geo.listGeofences();
+			expect(results.entries.length).toEqual(100);
+		});
+
+		test('listGeofences gets the second 100 geofences when nextToken is passed', async () => {
+			jest.spyOn(Credentials, 'get').mockImplementation(() => {
+				return Promise.resolve(credentials);
+			});
+
+			LocationClient.prototype.send = jest
+				.fn()
+				.mockImplementation(mockListGeofencesCommand);
+
+			const geo = new GeoClass();
+			geo.configure(awsConfig);
+
+			// Check that results are what's expected
+
+			const first100Geofences = await geo.listGeofences();
+
+			const second100Geofences = await geo.listGeofences({
+				nextToken: first100Geofences.nextToken,
+			});
+
+			expect(second100Geofences.entries.length).toEqual(100);
+			expect(second100Geofences.entries[0].geofenceId).toEqual(
+				'validGeofenceId100'
+			);
+			expect(second100Geofences.entries[99].geofenceId).toEqual(
+				'validGeofenceId199'
 			);
 		});
 	});