+++ packages/amplify-ui-storybook/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.40](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@2.0.39...@aws-amplify/ui-storybook@2.0.40) (2022-05-12)
+
+**Note:** Version bump only for package @aws-amplify/ui-storybook
+
+
+
+
+
+## [2.0.39](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@2.0.38...@aws-amplify/ui-storybook@2.0.39) (2022-05-03)
+
+**Note:** Version bump only for package @aws-amplify/ui-storybook
+
+
+
+
+
+## [2.0.38](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@2.0.37...@aws-amplify/ui-storybook@2.0.38) (2022-04-14)
+
+**Note:** Version bump only for package @aws-amplify/ui-storybook
+
+
+
+
+
+## [2.0.37](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@2.0.36...@aws-amplify/ui-storybook@2.0.37) (2022-04-04)
+
+**Note:** Version bump only for package @aws-amplify/ui-storybook
+
+
+
+
+
+## [2.0.36](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@2.0.35...@aws-amplify/ui-storybook@2.0.36) (2022-03-28)
+
+**Note:** Version bump only for package @aws-amplify/ui-storybook
+
+
+
+
+
+## [2.0.35](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@2.0.34...@aws-amplify/ui-storybook@2.0.35) (2022-03-22)
+
+**Note:** Version bump only for package @aws-amplify/ui-storybook
+
+
+
+
+
+## [2.0.34](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@2.0.33...@aws-amplify/ui-storybook@2.0.34) (2022-03-10)
+
+**Note:** Version bump only for package @aws-amplify/ui-storybook
+
+
+
+
+
 ## [2.0.33](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-storybook@2.0.32...@aws-amplify/ui-storybook@2.0.33) (2022-02-28)
 
 **Note:** Version bump only for package @aws-amplify/ui-storybook