+++ packages/analytics/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.6...@aws-amplify/analytics@5.2.7) (2022-05-12)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.2.6](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.5...@aws-amplify/analytics@5.2.6) (2022-05-03)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.2.5](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.4...@aws-amplify/analytics@5.2.5) (2022-04-14)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.2.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.3...@aws-amplify/analytics@5.2.4) (2022-04-04)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.2.3](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.2...@aws-amplify/analytics@5.2.3) (2022-03-28)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.2.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.1...@aws-amplify/analytics@5.2.2) (2022-03-22)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.2.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.0...@aws-amplify/analytics@5.2.1) (2022-03-10)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
 # [5.2.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.1.12...@aws-amplify/analytics@5.2.0) (2022-02-28)