+++ packages/pubsub/CHANGELOG.md
@@ -3,6 +3,75 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.4.0...@aws-amplify/pubsub@4.4.1) (2022-05-12)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+
+
+
+
+# [4.4.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.3.2...@aws-amplify/pubsub@4.4.0) (2022-05-03)
+
+
+### Features
+
+* **pubsub:** Add test coverage for the AWSAppSyncRealTimeProvider ([#9778](https://github.com/aws-amplify/amplify-js/issues/9778)) ([348366a](https://github.com/aws-amplify/amplify-js/commit/348366a044be2a3364c956f3a59ea125a7fb7d58))
+
+
+
+
+
+## [4.3.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.3.1...@aws-amplify/pubsub@4.3.2) (2022-04-14)
+
+
+### Bug Fixes
+
+* PubSub - Resolve ts strict errors in the AWSAppSyncRealTimeProvider ([#9731](https://github.com/aws-amplify/amplify-js/issues/9731)) ([0c9c401](https://github.com/aws-amplify/amplify-js/commit/0c9c4018573df31f7e4739207be22eefb8bf8a56))
+* PubSub - Resolve ts strict errors in the MqttOverWSProvider ([#9732](https://github.com/aws-amplify/amplify-js/issues/9732)) ([f50bd68](https://github.com/aws-amplify/amplify-js/commit/f50bd68b8e528c59d8e81b8978fa85a68a3120c1))
+
+
+
+
+
+## [4.3.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.3.0...@aws-amplify/pubsub@4.3.1) (2022-04-04)
+
+
+### Bug Fixes
+
+* PubSub - Resolve ts strict errors in the root import object ([#9733](https://github.com/aws-amplify/amplify-js/issues/9733)) ([c1d360c](https://github.com/aws-amplify/amplify-js/commit/c1d360c4acb388afb7b40cf73ba0484b6cdb64b1))
+
+
+
+
+
+# [4.3.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.2.11...@aws-amplify/pubsub@4.3.0) (2022-03-28)
+
+
+### Features
+
+* PubSub Add Options objects for all Providers and fix: Spelling error and deprecation of old exports ([#9683](https://github.com/aws-amplify/amplify-js/issues/9683)) ([b535af2](https://github.com/aws-amplify/amplify-js/commit/b535af2133b5460c6e8e2fcfd89e1fe235872c27))
+
+
+
+
+
+## [4.2.11](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.2.10...@aws-amplify/pubsub@4.2.11) (2022-03-22)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+
+
+
+
+## [4.2.10](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.2.9...@aws-amplify/pubsub@4.2.10) (2022-03-10)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+
+
+
+
 ## [4.2.9](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.2.8...@aws-amplify/pubsub@4.2.9) (2022-02-28)
 
 **Note:** Version bump only for package @aws-amplify/pubsub